requirements.txt
requests
beautifulsoup4
pandas
snowflake-connector-python[pandas]
python-dotenv
streamlit